Refactor for modularizing code
